refactor(code-smells): remove code-smells; set error-level to 4

+++ tests/integration/Models/MetaTest.php
@@ -24,7 +24,7 @@ class MetaTest extends TestCaseAuthenticated
     {
         $this->checkIfTestShouldBeSkipped("EVENT_AGENDA_SHOULD_TEST");
 
-        $eventId = TestData::getValue("EVENT_AGENDA_EVENT_ID");
+        $eventId = (int)TestData::getValue("EVENT_AGENDA_EVENT_ID");
         $agenda = EventAgendaRequest::fromEvent($eventId)->get();
 
         $this->assertModelHasValidMeta($agenda);
